+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Based on the suggestion at https://keepachangelog.com/en/1.0.0/.
 ## [Unreleased]
 - Upgraded NPM packages with `npm upgrade`.
 - Added a "demo mode", which loads and reasons over the example file.
+- Adds support for retrieving the list of all species within the clade
+  from the Open Tree Resolver, and then to retrieve the list of GBIF IDs
+  for those species.
 
 ## [0.2.0] - 2020-05-19
 - Added support for gzipping the ontology before upload to the reasoner.